+++ iftttie/channels/nest.py
@@ -75,6 +75,7 @@ def yield_events(event: MessageEvent) -> Iterable[Event]:
                 unit=Unit.IMAGE_URL,
                 timestamp=datetime.strptime(last_event['start_time'], timestamp_format),
                 title=f'{camera["name"]} Last Event',
+                is_logged=False,
             )
 
 
@@ -86,4 +87,5 @@ def yield_devices_events(devices: dict, prefix: str, keys: List[Tuple[str, Unit,
                 value=device[key],
                 unit=unit,
                 title=f'{device["name"]} {title}',
+                is_logged=(unit != Unit.IMAGE_URL),
             )