+++ iftttie/channels/base.py
@@ -6,6 +6,7 @@ from typing import Any
 from iftttie import web
 
 
+# TODO: rename to service?
 class BaseChannel(metaclass=ABCMeta):
     @abstractmethod
     async def run(self, context: web.Context, **kwargs: Any):