+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## `master`
 
-- **Change: #54 `Event.key` is changed to `Event.channel_id`** 
+- **Change: #54 `Event.key` is changed to `Event.channel_id`**
 - New: #17 automatically refresh page every minute
 - Fix: #64 Nest image URLs should not be logged
 - Chore: do not subclass `aiohttp.web.Application`
+- Chore: #54 refactor channel runner
 
 ## `0.10.0`