+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - New: #17 automatically refresh page every minute
 - New: #66 print raw event data on the channel page
 - Fix: #64 Nest image URLs should not be logged
+- Change: brush up the channel page
 - Chore: do not subclass `aiohttp.web.Application`
 - Chore: #54 refactor channel runner