+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## `master`
+## `0.11.0`
 
 - **Change: #54 `Event.key` is changed to `Event.channel_id`**
 - New: #17 automatically refresh page every minute